enhance server side latency

+++ services/deleteService.js
@@ -1,21 +1,14 @@
 import supabase from "./supabase.js";
 
-export const deleteJournalImageService = async(token, filepath) =>{
-    if(!token){
-        console.error('token is undefined');
-        throw {status: 400, error:'token is undefined'};
+export const deleteJournalImageService = async(userId, filepath) =>{
+    if(!userId){
+        console.error('userId is undefined');
+        throw {status: 400, error:'userId is undefined'};
     }
     
     if(!filepath){
         console.error('filepath is undefined');
         throw {status: 400, error: 'filepath is undefined'}
-    }
-
-    const {data: authData,error: errorAuthData} = await supabase.auth.getUser(token);
-
-    if(errorAuthData){
-        console.error('error while validating user:', errorAuthData.message);
-        throw {status: 500, error: 'user is not authorized'}
     }
 
     const {error} = await supabase.storage
@@ -30,21 +23,22 @@ export const deleteJournalImageService = async(token, filepath) =>{
     return true;
 }
 
-export const deleteJournalContentService = async(journalId, token) =>{
+export const deleteJournalContentService = async(journalId, userId) =>{
     if(!journalId){
         console.error('journalid is undefined')
         throw {status: 400, error:' journalId is undefined'}
     }
 
-    if(!token){
-        console.error('token is undefined');
-        throw {status: 400, error: 'token is undefined'}
+    if(!userId){
+        console.error('userId is undefined');
+        throw {status: 400, error: 'userId is undefined'}
     }
 
     const {error: errorDeletingData} = await supabase
     .from('journals')
     .delete()
     .eq('id', journalId)
+    .eq('user_id', userId)
 
     if(errorDeletingData){
         console.error('supabase error while deleting the journal content', errorDeletingData.message);
